move react parser and tests to a separate folder
