+++ src/jsdoc/react/index.ts
@@ -1,0 +1,3 @@
+import parse from './react-parser';
+
+export default parse;
