+++ src/jsdoc/vue/index.ts
@@ -1,0 +1,3 @@
+import parse from './vue-parser';
+
+export default parse;
