+++ src/jsdoc/vue/vue-parser.ts
@@ -1,9 +1,9 @@
 import * as domain from 'domain';
 import vuedoc from '@vuedoc/parser';
-import { PathOsBased } from '../utils/path';
-import { Doclet } from './parser';
-import { pathNormalizeToLinux } from '../utils';
-import logger from '../logger/logger';
+import { PathOsBased } from '../../utils/path';
+import { Doclet } from '../parser';
+import { pathNormalizeToLinux } from '../../utils';
+import logger from '../../logger/logger';
 
 function formatProperty(prop) {
   const { name, description, type, required } = prop;
